refactored passport-google config to sit in passport-config
